Reorganize delivery-route components.
