+++ client/src/app/modules/auth/modules/employee/employee-routing.module.ts
@@ -19,8 +19,8 @@ import { ClientsComponent } from './components/clients/clients/clients.component
 import { ClientDetailComponent } from './components/clients/client-detail/client-detail.component'; 
 import { UsersComponent } from './components/users/users.component';
 import { UserDetailComponent } from './components/user-detail/user-detail.component';
-import { DeliveryRoutesComponent } from './components/delivery-routes/delivery-routes.component';
-import { DeliveryRouteDetailComponent } from './components/delivery-route-detail/delivery-route-detail.component';
+import { DeliveryRoutesComponent } from './components/delivery-routes/delivery-routes/delivery-routes.component';
+import { DeliveryRouteDetailComponent } from './components/delivery-routes/delivery-route-detail/delivery-route-detail.component';
 import { ProductsComponent } from './components/products/products.component';
 import { ProductDetailComponent } from './components/product-detail/product-detail.component';
